feat: Implement PSE payment method with bank selection
- Added bank selection dropdown in AlternativePaymentMethods component for PSE payments.
- Integrated fetchBanks utility to retrieve available banks for PSE.
- Updated PaymentForm to handle selected bank state and pass it to AlternativePaymentMethods.
- Created PsePaymentData type to structure PSE payment data.
- Implemented payWithPse utility function to handle PSE payment processing.
- Developed Step4OrderSummary component to display order summary and total.
- Introduced useCheckoutLogic hook to manage checkout state and payment processing logic.
- Enhanced card validation utility to ensure robust validation of card fields.
